Add warning to volunteer translators to README-translators.txt
923efe54d6cb5ff350bd401511a24ef0c484962b added a note to the PR
template, but the best place is probably in the translators readme.
Hopefully this will make sure we no longer get people eager to start
a translation because nobody else had started one yet, and then have to
be disappointed with a reply from us saying we can't accept voluntary
translation contributions.