Fix blurry images when starting the game filtered
If the game started in linear mode, the graphics would be created with
linear filtering on. This means, when you switch off of linear
filtering, all of the assets are still blurry.

This commit makes sure that the assets are loaded as nearest,
then it toggles linear on only after.